feat(games): Fixed and added some game settings
Games are now launchable from anywhere in the application via the coin
slot button (spacebar). also added a creators folder for the creation
scripts.
Made a change to the scripts so they work in a new directory.
Added an active flag in the game manifest so inprogress games are not
loaded (safety feature)